Rename download button component
