+++ resources/js/components/shared/index.js
@@ -20,7 +20,7 @@ import Required from "./Required";
 import SidebarButton from "./SidebarButton";
 import SidebarNav from "./SidebarNav";
 import SliderWithInput from "./SliderWithInput";
-import SvgToPdfButton from "./SvgToPdfButton.vue";
+import DownloadSvgButton from "./DownloadSvgButton.vue";
 
 export {
   BasicSearch,
@@ -45,5 +45,5 @@ export {
   SidebarButton,
   SidebarNav,
   SliderWithInput,
-  SvgToPdfButton,
+  DownloadSvgButton,
 };
